Update convention document magic link lifetime from "long" to "short"

+++ back/src/domains/convention/use-cases/notifications/NotifyAllActorsOfFinalConventionValidation.ts
@@ -168,7 +168,7 @@ export class NotifyAllActorsOfFinalConventionValidation extends TransactionalUse
         )}`
       : await makeShortMagicLink({
           targetRoute: frontRoutes.conventionDocument,
-          lifetime: "long",
+          lifetime: "short",
           singleUse: false,
         });
